LaserOptimizerSuite/requirements.txt
numpy
matplotlib
scipy
svg.path
